move deprecated test
